third_party/blink/web_tests/platform/generic/media/media-fragments/TC0080-expected.txt


Title: TC0080
Fragment: 't=00:00.'
Comment: A valid media fragment: { starting at 0 seconds. Thus, the UA will play the entire media resource.
EVENT(canplaythrough)
EXPECTED (video.currentTime == '0') OK
RUN(video.play())
EVENT(pause)
EXPECTED (video.currentTime - fragmentEndTime <= '0.5') OK
END OF TEST
